Add icon flag to bake command
This change adds a required `--icon` flag to the bake command. The flag
takes a path to an icon file, which will then be base64 encoded in the
tile's metadata as an `icon_image`. For now we do not error on an
`icon_image` being present in the `base.yml`.

[finishes #152909373]

Signed-off-by: James Myers <jmyers@pivotal.io>