Refactor and start to use the regex
- Use Decoder instead of Unmarhshalling
- Use var variable-name type when initializing defaults
- Pull all items in the bucket and find those that match the regex

TODO:

Store the matching S3 objects into a new struct that saves the path (aka
key), name, version, stemcell, stemcell version, etc. And add this to a
new object store of MatchedS3Objects

Iterate through the `assets.lock` checking to see if the releases listed
there are actually in the MatchedS3Objects list, if not, error.

If it's found, we have the path and can construct the download

[#160557336] null

Signed-off-by: Caitlyn Yu <cyu@pivotal.io>