Make the checksum optional in assets.lock
This is a quality of life feature for releases on bosh.io.
Without this change, verifying the checksum will fail because it can't
find the release file (the name is different than when pulled from s3).

Even if we resolved the issue with finding the file, the user would have
to pre-download the release and checksum it to give a correct value.

[#166336627] kiln fetch can fetch a tarball of a release if it is on bosh.io

Authored-by: Matt Royal <mroyal@pivotal.io>