assets.yml can configure any number of release sources
Previously there was 1 compiled release source, 1 uncompiled release
source, and an implicit bosh.io release source.

Now all of those release source types still exist, but more than one of
each can be defined, and they can be specified in any order. This
requires a change to the format of assets.yml. See changes to the
README.md for details on the new format

**THIS IS A BREAKING CHANGE**

Authored-by: Matt Royal <mroyal@pivotal.io>

[#168937284] kiln fetch can fetch built releases from the s3 bucket used by delivery CI