Allow file paths for publish to be absolute or relative paths
The argument passed to osfs.New() determines whether to limit files to a
specific directory or not. By making it "." we were sandboxing the user
into the current directory. We removed this, which means they can
reference any file on the file system now.

We couldn't think of a way to test this outside of the acceptance tests.
We decided not to add a new acceptance tests, since those are
destructive, slow, and more likely to fail spuriously.

Co-authored-by: Matt Royal <mroyal@pivotal.io>
Co-authored-by: Gary Liu <galiu@pivotal.io>

[#168547537] `kiln publish` should take absolute paths instead of relative