kiln bake always accepts --kilnfile
Previously --kilnfile was mutually exclusive with either of the
--stemcell* flags. Instead, the --stemcell* flags now override the
stemcell source for bake and the value in the kilnfile is ignored

One behavior change results from this:
Previously you could omit the --stemcell* flags AND the --kilnfile flag.
There was an acceptance test for that case, but I don't believe it was
realistic. In practice, I believe that use case would have errored out
in the interpolation stage, since all tile repos use the `stemcell`
function in their templates, and this requires that at least one
stemcell manifest is provided.

Authored-by: Matt Royal <mroyal@pivotal.io>

[#169855792] Make `--kilnfile` a "global flag" in `kiln`