Output sooner so the user knows we're working
If you're not logged in to lpass properly, the standard way we invoke
kiln fetch will hang indefinitely. (We don't know how to fix this because
it's a side effect of how *nix pipes work.) The "DownloadRelease" command also
takes a minute to start producing output, so it's hard to tell if you're
seeing normal behavior or the "not logged in" problem.

Printing a bit right after we've parsed flags but before we start the
long operation makes it clear that all is well.

Co-authored-by: Nat Bennett <nbennett@pivotal.io>
Co-authored-by: Matt Royal <mroyal@pivotal.io>

[#170218474] Pipestat: Review and close tasks in bugs/incidents in the delivery tracker