Push upload logic into S3BuiltReleaseSource
The previous code in UploadRelease was very tightly coupled to the
specifics of the release source definition and the way the
S3BuiltReleaseSource interacts with S3. By extracting this into the
UploadRelease method, we've returned to a structure where only
S3BuiltReleaseSource needs to know these details.

I chose to repurpose the ReleaseSourceFactory, but then use type
coercion to find sources that can also be used for upload. I'm not sure
this is the best solution: a dedicated ReleaseUploaderFactory may be a
better solution to this problem.

Authored-by: Matt Royal <mroyal@pivotal.io>

[#170681713] kiln upload-release errors when upload would replace existing file