Use release, version and SHA1 to validate release files
LocalReleaseDirectory no longer knows anything about built or compiled
releases. Instead, it returns a generic release.Local that has a name,
version and SHA1. `kiln fetch` now compares the SHA1 of the local file
to the one specified in the Kilnfile.lock, and deletes any files that
don't match.

This also simplified the following places:
- `kiln update-release` now just pulls the SHA1 from local release
  instead of calculating it

- `kiln fetch` now validates the SHA1 of each file immediately after
  downloading it, instead of checking all SHAs via LocalReleaseDirectory
  as a final step

Authored-by: Matt Royal <mroyal@pivotal.io>