Add concept of a ReleaseUploaderFactory
When you're uploading a release, you want to upload to a specific
source. Due to this, it doesn't make sense for the MultiReleaseSource to
handle uploading. Instead, we now use a ReleaseUploaderFactory, which
takes the release source name as an argument and returns the specific
release source that was requested (or an error if it wasn't found)

Authored-by: Matt Royal <mroyal@pivotal.io>