Copy the releaseLocks slice before mutating it
The current code doesn't use the slice again, so it wasn't a problem
that we modified the slice in place. However, if future code had tried to
read the value then weird bugs would have occurred. This is because the
previous code was mutating the slice that lives on the KilnfileLock
struct.

Authored-by: Matt Royal <mroyal@pivotal.io>