WIP - Making an attempt to port help
The current problem is that we always try to parse the Kilnfile and
Kilnfile.lock, whether or not we need them.

One possible solution: instead of having Dependencies be a struct, give
it memoized methods for retrieving Kilnfile and Kilnfile.lock, etc. That
way we won't try to load the files unless we need them

Authored-by: Matt Royal <mroyal@pivotal.io>

[#171270560] kiln upload-release fails when given a compiled release