Only skip updating entry when --skip-same-version is passed
We need to update the remote and SHA1 every time for now, because the
same release will suddenly have a different remote_path and sha1 when we
do component freeze and cut a new pipeline. This is due to the fact that
we keep different copies of the releases for each PAS version

For example:
  built-releases/2.10/diego/diego-1.2.3.tgz
    vs
  built-releases/2.9/diego/diego-1.2.3.tgz

  Each has a unique path and a unique SHA1

This will no longer be true once we move over to using bosh.io and
final-pcf-bosh-releases as our only sources for built releases. At that
point we can start using --skip-same-version to let us "autobump" based
on built releases, but only downgrade from the compiled version when the
version actually changes.

Authored-by: Matt Royal <mroyal@pivotal.io>

[#171456987] kiln sync-with-local only updates an item if the version has changed