Fix compilation failure on darwin caused by dependency bumps
It looks like compilation failed on MacOS due to some
compilation error in some dependency. Instead of digging
into it, I did 'go get -u ./...' and ran tests and compiled
the tool and it all works now.

I also set GOOS and GOARCH locally to cross compile to
see if it works before commiting this and it all works
now.

Authored-by: Christopher Hunter <hunterch@vmware.com>