set useful defaults for Bake flags
Most of our "bake" invocations require 7+ flags.
The readme for Kiln suggests that Kiln is opinionated.
In this commit we add some week opinions about directory
and file naming.

Existing invocations of the "bake" command _should_ continue
to work similar to how they did before this change; no
(existing) tests were harmed (modified) in the making of this
feature.

The new behavior uses default values iff os.Stat does NOT
return an error when passed the default value.

  if a flag(s) are set,
    use them without modification
  else if the default value is a valid path
    use it
  else
    the option gets it's zero value
      for variatic flags this is a len=0 slice
      for strings (paths to files) this is the empty string

The example tile "./build" script takes advantage of the
new defaults but also uses some non-defaulted options.

We changed the paths in the example-tile to match our
internal conventions on the Platform Provider Experience
team. The defaults are also based on those conventions.

We decided against adding directory and path configuration
to the Kilnfile as it would make implementation more
complicated and may make it easier for users to ignore
our opinions. We believe adding useful defaults and allowing
overrides via flags rewards conformity but still allows
divergence.

Co-authored-by: Claire Tinati <ctinati@vmware.com>