standardize kilnfile loading and parsing across commands
This change uses the flag parsing from
https://github.com/pivotal-cf/kiln/pull/257 to reduce the number of
flags requred for commands by loading reasonable defaults if they
make sense (the expected default exists relative to the Kilnfile).

I used this change to move the flag parsing from the cargo package
to a flags package. This is intended as a step towards cargo being
a place for kiln related data structures and procedures.