feat (cach-compiled-releases): don't download bad releases
BOSH will allow you to export and download a release without compiled packages.
This happens when you upload a compiled release and try to export the release for a different stemcell.
This is bad because the exported release tarball is invalid.

Big thanks to Brian C, Joseph P, Mark S, and Steve T for finding this issue!

Story: #181964258