test: implement component update tests
This drove out a refactor of the test setup.
The change makes it easier to execute just one feature.
It also ensures the test name matches the feature file name.

In order to test the standard output of find-release-version, there is a new universal setup where we add an additional io.Writer to stdout and stderr before calling them.
This allows the exec step func to read out the output and check for a substring.

I added a shared helper function to setup a github client.
It is also used for the release-notes command.

I realized the steps package name was not great because each of the Initialize* functions also did before/after scenario stuff.
So I renamed the package back to what we had it named before.

I renamed a few feature files so they would be more general and it would be more intuative to add additional scenarios to them.