refactor: function and method names in scenario package
I also renamed the "Environment" identifier to "opsManagerEnvironment".

I made the shared state helper functions smaller and more focused on managing state. The "business logic" was moved to either a separate file (ops_manager_environment.go or utilities.go).

I also made the ScenarioContext wrapper a private thing. It does not need to be part of the public API of this package.