Feat: Generate-OSM-Manifest
[Exceptional OSM for Foundation OSS in TAS
4.0](https://3.basecamp.com/4957863/projects/31968079)

This command was developed to fill the requirements of the bet above.

Generate-OSM-Manifest will generate an OSM-compatible manifest for an
operator, specifying only a Kilnfile and a Github Access Token.

Unless specified by running `kiln generate-osm-manifest --no-download`,
this command will also download and compress each release that is within
the specified Kilnfile.

Co-Authored-By: Ryan Hall <hallr@vmware.com>
Co-Authored-By: David Timm <dtimm@vmware.com>