refactor: move tile test code into a separate package
By moving the tile test code to a separate directory, the command code can focus on flag parsing and user experience.
The tile package can be freely factored to improve understanding of the container interactions.

This change also fixes some conncurrency bugs that led to leaked go-routines.

It should have nearly identical behavior from an end user perspective with a few small changes:
- some short flags were removed (multi character short flags are weird, if a single character does not make sense, remove it).

Also the test of a failing tile test was removed. This could easily be added back in.
A unit test covers a similar scenario and the integration test in the test package only tests the success case.

If you run 'go test -short ./...' from the root of the project the test should run in a few seconds (rather than a few minutes).