feat: make find-release-version output match Kilnfile.lock
We from cryo just started using kiln for our tile builds. We have quite
a few tiles and mostly want to bump versions in a tile to latest since
we don't ship many releases in a tile.
I found the output of the find-release-version command somewhat
confusing because it doesn't really match the structure of the
Kilnfile.lock so I figured I'd try to PR this in.

What does this do? It drops the `releaseVersionOutput` that wasn't used
anywhere else and just uses `BOSHReleaseTarballLock` to generate the
output.

It adds the option to output yaml (in case you want to use the output
and put it into the lockfile..)

it also updates the `BOSHReleaseTarballLock` with json tags so
marshalling produces the expected output.

Authored-by: Konstantin Kiess <kkiess@vmware.com>