add alternative to execute image-build every time
why?:

because some CI systems cannot get past the dockerhub rate limitation.
We want to use kiln test for the csb tiles. But we use the tpe / runway
concourse. These instances make it impossible to communicate with dockerhub
because of rate limits ( even with authenticated pulls )

We can make the `kiln test` execution work by pre loading the images that
are used as FROM stages in `internal/test/Dockerfile` into the docker host
we start in the concourse task. Essentially we end up pulling the images from
an accessible repository, then we retag them to match the expected FROM args
and that makes the implicit image build work.

Instead it would be nice to avoid having to run the build stage and make kiln
test consume a provided image instead so it can be run offline.